refactor: move api executor request/response into ../models
